Move ConnectedUsbDevicesFetcher to usb package.
